Architecture tidy up: 1.Augmentor adding responsible for all general training data augmentation 2.PreProcessor adding responsible for training data preprocessing, each model with one preprocessor 3.PostProcessor adding responsible for inference output postprocessing, each model with one postprocessor
