+++ hyperpose/Model/human.py
@@ -50,7 +50,17 @@ class Human:
     def get_area(self):
         bbx=self.get_bbx()
         self.area=float(bbx[2]*bbx[3])
-        return self.area 
+        return self.area
+    
+    def bias(self,bias_w,bias_h):
+        for part_idx in self.body_parts.keys():
+            body_part=self.body_parts[part_idx]
+            body_part.x=body_part.x+bias_w
+            body_part.y=body_part.y+bias_h
+            if(body_part.x<0):
+                body_part.x=-1000.0
+            if(body_part.y<0):
+                body_part.y=-1000.0
     
     def scale(self,scale_w,scale_h):
         for part_idx in self.body_parts.keys():
@@ -59,6 +69,11 @@ class Human:
             body_part.y=body_part.y*scale_h
             body_part.w=body_part.w*scale_w
             body_part.h=body_part.h*scale_h
+            if(body_part.x<0):
+                body_part.x=-1000.0
+            if(body_part.y<0):
+                body_part.y=-1000.0
+                
     
     def draw_human(self,img):
         img_h,img_w,img_c=img.shape
@@ -68,20 +83,22 @@ class Human:
             body_part=self.body_parts[part_idx]
             x=body_part.x
             y=body_part.y
+            if(x<0 or x>=img_w or y<0 or y>=img_h):
+                continue
             color=self.colors[part_idx]
             img=cv2.circle(img,(int(x),int(y)),radius=radius,color=color,thickness=-1)
         line_color=(255,0,0)
         for limb in self.limbs:
             src_part_idx,dst_part_idx=limb
-            if(src_part_idx==2 and dst_part_idx==16):
-                continue
-            elif(src_part_idx==5 and dst_part_idx==17):
-                continue
             if((src_part_idx in self.body_parts) and (dst_part_idx in self.body_parts)):
                 src_body_part=self.body_parts[src_part_idx]
                 src_x,src_y=int(src_body_part.x),int(src_body_part.y)
                 dst_body_part=self.body_parts[dst_part_idx]
                 dst_x,dst_y=int(dst_body_part.x),int(dst_body_part.y)
+                if(src_x<0 or src_x>=img_w or src_y<0 or src_y>=img_h):
+                    continue
+                if(dst_x<0 or dst_x>=img_w or dst_y<0 or dst_y>=img_h):
+                    continue
                 img=cv2.line(img,(src_x,src_y),(dst_x,dst_y),color=line_color,thickness=thickness)
         return img
