+++ hyperpose/Model/pifpaf/eval.py
@@ -7,75 +7,94 @@ import scipy.stats as st
 from functools import partial
 import multiprocessing
 import matplotlib.pyplot as plt
-from .infer import Post_Processor
-from .utils import get_hr_conf,get_arrow_map
+from .processor import PostProcessor
+from .utils import get_hr_conf,get_arrow_map,maps_to_numpy
+from ..common import pad_image_shape
 
-def infer_one_img(model,post_processor,img,image_id=-1,is_visual=False,save_dir="./save_dir"):
+def infer_one_img(model,postprocessor,img,img_id=-1,is_visual=False,save_dir="./save_dir",enable_multiscale_search=False,debug=True):
     img=img.numpy().astype(np.float32)
     img_id=img_id.numpy()
+    if(debug):
+        print(f"infer image id:{img_id}")
     #TODO: use padded-scale that wouldn't casue deformation
     img_h,img_w,_=img.shape
-    input_image=cv2.resize(img,(model.win,model.hin))[np.newaxis,:,:,:]
+    hin,win=model.hin,model.win
+    scale_rate=min(hin/img_h,win/img_w)*0.95
+    scale_h,scale_w=int(scale_rate*img_h),int(scale_rate*img_w)
+    scale_image=cv2.resize(img,(scale_w,scale_h),interpolation=cv2.INTER_CUBIC)
+    padded_image,pad=pad_image_shape(scale_image,shape=(hin,win),pad_value=0.0)
     #default channels_first
-    input_image=np.transpose(input_image,[0,3,1,2])
-    pif_maps,paf_maps=model.forward(input_image,is_train=False)
-    pif_maps=[pif_map[0] for pif_map in pif_maps]
-    paf_maps=[paf_map[0] for paf_map in paf_maps]
-    ret_humans=post_processor.process(pif_maps,paf_maps)
+    input_image=np.transpose(padded_image[np.newaxis,:,:,:].astype(np.float32),[0,3,1,2])
+    pif_maps,paf_maps=model.infer(input_image)
+    pif_maps=maps_to_numpy([pif_map[0] for pif_map in pif_maps])
+    paf_maps=maps_to_numpy([paf_map[0] for paf_map in paf_maps])
+    ret_humans=postprocessor.process(pif_maps,paf_maps)
+    for ret_human in ret_humans:
+        ret_human.bias(bias_w=-pad[2],bias_h=-pad[0])
+        ret_human.scale(scale_w=1/scale_rate,scale_h=1/scale_rate)
     if(is_visual):
-        visualize(img,img_id,pif_maps,paf_maps,ret_humans,stride=post_processor.stride,save_dir=save_dir)
+        visualize(img,img_id,padded_image,pif_maps,paf_maps,ret_humans,stride=post_processor.stride,save_dir=save_dir)
     return ret_humans
 
-def visualize(img,img_id,pd_pif_maps,pd_paf_maps,humans,stride=8,save_dir=".save_dir"):
+def visualize(img,img_id,processed_img,pd_pif_maps,pd_paf_maps,humans,stride=8,save_dir="./save_dir"):
     print(f"{len(humans)} human found!")
     print("visualizing...")
     os.makedirs(save_dir,exist_ok=True)
     ori_img=np.clip(img*255.0,0.0,255.0).astype(np.uint8)
+    processed_img=np.clip(processed_img*255.0,0.0,255.0).astype(np.uint8)
     #get ouput_result
     vis_img=ori_img.copy()
     for human in humans:
         vis_img=human.draw_human(vis_img)
     #decode result_maps
-    stride=model.shape
-    pd_pif_conf,pd_pif_vec,_,pd_pif_scale=pd_pif_maps
-    pd_paf_conf,pd_paf_src_vec,pd_paf_dst_vec,_,_,_,_=pd_paf_maps
+    pd_pif_conf,pd_pif_vec,pd_pif_scale=pd_pif_maps
+    pd_paf_conf,pd_paf_src_vec,pd_paf_dst_vec,_,_,=pd_paf_maps
     pd_pif_conf_show=np.amax(pd_pif_conf,axis=0)
     pd_pif_hr_conf_show=np.amax(get_hr_conf(pd_pif_conf,pd_pif_vec,pd_pif_scale,stride=stride,thresh=0.1),axis=0)
     pd_paf_conf_show=np.amax(pd_paf_conf,axis=0)
-    pd_paf_vec_show=np.zeros(pd_pif_hr_conf_show.shape[0],pd_pif_hr_conf_show.shape[1],3)
+    pd_paf_vec_show=np.zeros(shape=(pd_pif_hr_conf_show.shape[0],pd_pif_hr_conf_show.shape[1],3)).astype(np.int8)
     pd_paf_vec_show=get_arrow_map(pd_paf_vec_show,pd_paf_conf,pd_paf_src_vec,pd_paf_dst_vec,thresh=0.1)
     #plt draw
     fig=plt.figure(figsize=(12,12))
     #show input image
-    a=fig.add_subplot(2,3,1)
+    a=fig.add_subplot(3,3,1)
     a.set_title("input image")
     plt.imshow(ori_img)
     #show output result
-    a=fig.add_subplot(2,3,4)
+    a=fig.add_subplot(3,3,3)
     a.set_title("output result")
     plt.imshow(vis_img)
+    #pif
+    #show processed image
+    a=fig.add_subplot(3,3,4)
+    a.set_title("processed image")
+    plt.imshow(processed_img)
     #show pif_conf_map
-    a=fig.add_subplot(2,3,2)
+    a=fig.add_subplot(3,3,5)
     a.set_title("pif_conf_map")
     plt.imshow(pd_pif_conf_show,alpha=0.8)
     plt.colorbar()
     #show pif_hr_conf_map
-    a=fig.add_subplot(2,3,3)
+    a=fig.add_subplot(3,3,6)
     a.set_title("pif_hr_conf_map")
     plt.imshow(pd_pif_hr_conf_show,alpha=0.8)
     plt.colorbar()
+    #paf
+    a=fig.add_subplot(3,3,7)
+    a.set_title("processed image")
+    plt.imshow(processed_img)
     #show paf_conf_map
-    a=fig.add_subplot(2,3,5)
+    a=fig.add_subplot(3,3,8)
     a.set_title("paf_conf_map")
     plt.imshow(pd_paf_conf_show,alpha=0.8)
     plt.colorbar()
     #show paf_vec_map
-    a=fig.add_subplot(2,3,6)
+    a=fig.add_subplot(3,3,9)
     a.set_title("paf_vec_map")
     plt.imshow(pd_paf_vec_show,alpha=0.8)
     plt.colorbar()
     #save fig
-    plt.savefig(os.path.join(save_dir,f"{image_id}_visualize.png"))
+    plt.savefig(os.path.join(save_dir,f"{img_id}_visualize.png"))
     plt.close()
 
 def _map_fn(image_file,image_id,hin,win):
@@ -119,18 +138,18 @@ def evaluate(model,dataset,config,vis_num=30,total_eval_num=10000,enable_multisc
     pd_anns=[]
     vis_dir=config.eval.vis_dir
     kpt_converter=dataset.get_output_kpt_cvter()
-    post_processor=Post_Processor(parts=model.parts,limbs=model.limbs,colors=model.colors)
+    postprocessor=PostProcessor(parts=model.parts,limbs=model.limbs,colors=model.colors,hin=model.hin,win=model.win,\
+        hout=model.hout,wout=model.wout,debug=False)
     
     eval_dataset=dataset.get_eval_dataset()
     paramed_map_fn=partial(_map_fn,hin=model.hin,win=model.win)
     eval_dataset=eval_dataset.map(paramed_map_fn,num_parallel_calls=max(multiprocessing.cpu_count()//2,1))
+    dataset_lenth=len(list(eval_dataset))
     for eval_num,(img,img_id) in enumerate(eval_dataset):
         if(eval_num>=total_eval_num):
             break
-        if(eval_num<=vis_num):
-            humans=infer_one_img(model,post_processor,img,img_id=img_id,is_visual=True,save_dir=vis_dir,enable_multiscale_search=enable_multiscale_search)
-        else:
-            humans=infer_one_img(model,post_processor,img,img_id=img_id,is_visual=False,save_dir=vis_dir,enable_multiscale_search=enable_multiscale_search)
+        is_visual=(eval_num<=vis_num)
+        humans=infer_one_img(model,postprocessor,img,img_id=img_id,is_visual=is_visual,save_dir=vis_dir,enable_multiscale_search=enable_multiscale_search)
         for human in humans:
             ann={}
             ann["category_id"]=1
@@ -141,14 +160,14 @@ def evaluate(model,dataset,config,vis_num=30,total_eval_num=10000,enable_multisc
             kpt_list=[]
             for part_idx in range(0,model.n_pos):
                 if(part_idx not in human.body_parts):
-                    kpt_list.append([-1000,-1000])
+                    kpt_list.append([-1000.0,-1000.0])
                 else:
                     body_part=human.body_parts[part_idx]
                     kpt_list.append([body_part.get_x(),body_part.get_y()])
             ann["keypoints"]=kpt_converter(kpt_list)
             pd_anns.append(ann)   
         if(eval_num%100==0):
-            print(f"evaluating {eval_num}/{len(list(eval_dataset))}")
-            
+            print(f"evaluating {eval_num}/{dataset_lenth}")
+    
     result_dic={"annotations":pd_anns}
     dataset.official_eval(result_dic,vis_dir)
