+++ hyperpose/Model/pifpaf/model.py
@@ -4,10 +4,12 @@ import tensorlayer as tl
 from tensorlayer import layers
 from tensorlayer.models import Model
 from tensorlayer.layers import BatchNorm2d, Conv2d, DepthwiseConv2d, LayerList, MaxPool2d
+from .define import CocoColor
 from ..backbones import Resnet50_backbone
 
+
 class Pifpaf(Model):
-    def __init__(self,parts,limbs,n_pos=17,n_limbs=19,hin=368,win=368,scale_size=16,backbone=None,pretraining=False,quad_size=2,quad_num=1,
+    def __init__(self,parts,limbs,colors=CocoColor,n_pos=17,n_limbs=19,hin=368,win=368,scale_size=32,backbone=None,pretraining=False,quad_size=2,quad_num=1,
     lambda_pif_conf=30.0,lambda_pif_vec=2.0,lambda_pif_scale=2.0,lambda_paf_conf=50.0,lambda_paf_src_vec=3.0,lambda_paf_dst_vec=3.0,
     lambda_paf_src_scale=2.0,lambda_paf_dst_scale=2.0,data_format="channels_first"):
         super().__init__()
@@ -15,14 +17,13 @@ class Pifpaf(Model):
         self.limbs=limbs
         self.n_pos=n_pos
         self.n_limbs=n_limbs
+        self.colors=colors
         self.hin=hin
         self.win=win
         self.quad_size=quad_size
         self.quad_num=quad_num
-        self.stride=scale_size
-        self.scale_size=scale_size/(self.quad_size**self.quad_num)
-        self.hout=int(hin/self.scale_size)
-        self.wout=int(win/self.scale_size)
+        self.scale_size=scale_size
+        self.stride=int(self.scale_size/(self.quad_size**self.quad_num))
         self.lambda_pif_conf=lambda_pif_conf
         self.lambda_pif_vec=lambda_pif_vec
         self.lambda_pif_scale=lambda_pif_scale
@@ -33,18 +34,22 @@ class Pifpaf(Model):
         self.lambda_paf_dst_scale=lambda_paf_dst_scale
         self.data_format=data_format
         if(backbone==None):
-            self.backbone=Resnet50_backbone(data_format=data_format,use_pool=False,scale_size=32,decay=0.99,eps=1e-4)
+            self.backbone=Resnet50_backbone(data_format=data_format,use_pool=False,scale_size=self.scale_size,decay=0.99,eps=1e-4)
+            self.stride=int(self.stride/2) #because of not using max_pool layer of resnet50
         else:
             self.backbone=backbone(data_format=data_format,scale_size=self.scale_size)
-        self.pif_head=self.PifHead(input_features=self.backbone.out_channels,n_pos=self.n_pos,n_limbs=self.n_limbs,\
-            quad_size=self.quad_size,data_format=self.data_format)
-        self.paf_head=self.PafHead(input_features=self.backbone.out_channels,n_pos=self.n_pos,n_limbs=self.n_limbs,\
-            quad_size=self.quad_size,data_format=self.data_format)
+        self.hout=int(hin/self.stride)
+        self.wout=int(win/self.stride)
         #generate mesh grid
         x_range=np.linspace(start=0,stop=self.wout-1,num=self.wout)
         y_range=np.linspace(start=0,stop=self.hout-1,num=self.hout)
         mesh_x,mesh_y=np.meshgrid(x_range,y_range)
         self.mesh_grid=np.stack([mesh_x,mesh_y])
+        #construct head
+        self.pif_head=self.PifHead(input_features=self.backbone.out_channels,n_pos=self.n_pos,n_limbs=self.n_limbs,\
+            quad_size=self.quad_size,hout=self.hout,wout=self.wout,stride=self.stride,mesh_grid=self.mesh_grid,data_format=self.data_format)
+        self.paf_head=self.PafHead(input_features=self.backbone.out_channels,n_pos=self.n_pos,n_limbs=self.n_limbs,\
+            quad_size=self.quad_size,hout=self.hout,wout=self.wout,stride=self.stride,mesh_grid=self.mesh_grid,data_format=self.data_format)
     
     @tf.function(experimental_relax_shapes=True)
     def forward(self,x,is_train=False):
@@ -136,13 +141,17 @@ class Pifpaf(Model):
         return loss_pif_maps,loss_paf_maps,total_loss
     
     class PifHead(Model):
-        def __init__(self,input_features=2048,n_pos=19,n_limbs=19,quad_size=2,data_format="channels_first"):
+        def __init__(self,input_features=2048,n_pos=19,n_limbs=19,quad_size=2,hout=8,wout=8,stride=8,mesh_grid=None,data_format="channels_first"):
             super().__init__()
             self.input_features=input_features
             self.n_pos=n_pos
             self.n_limbs=n_limbs
+            self.hout=hout
+            self.wout=wout
+            self.stride=stride
             self.quad_size=quad_size
             self.out_features=self.n_pos*5*(self.quad_size**2)
+            self.mesh_grid=mesh_grid
             self.data_format=data_format
             self.tf_data_format="NCHW" if self.data_format=="channels_first" else "NHWC"
             self.main_block=Conv2d(n_filter=self.out_features,in_channels=self.input_features,filter_size=(1,1),data_format=self.data_format)
@@ -150,25 +159,31 @@ class Pifpaf(Model):
         def forward(self,x,is_train=False):
             x=self.main_block.forward(x)
             x=tf.nn.depth_to_space(x,block_size=self.quad_size,data_format=self.tf_data_format)
-            x=tf.reshape(x,[x.shape[0],self.n_pos,5,x.shape[2],x.shape[3]])
+            x=tf.reshape(x,[-1,self.n_pos,5,self.hout,self.wout])
             pif_conf=x[:,:,0,:,:]
             pif_vec=x[:,:,1:3,:,:]
             pif_logb=x[:,:,3,:,:]
             pif_scale=tf.exp(x[:,:,4,:,:])
             #restore vec_maps in inference
             if(is_train==False):
-                pif_conf=tf.nn.sigmoid(pif_conf)
-                pif_vec[:,:]+=self.mesh_grid
+                infer_pif_conf=tf.nn.sigmoid(pif_conf)
+                infer_pif_vec=(pif_vec[:,:]+self.mesh_grid)*self.stride
+                infer_pif_scale=pif_scale*self.stride
+                return infer_pif_conf,infer_pif_vec,pif_logb,infer_pif_scale
             return pif_conf,pif_vec,pif_logb,pif_scale
         
     class PafHead(Model):
-        def __init__(self,input_features=2048,n_pos=19,n_limbs=19,quad_size=2,data_format="channels_first"):
+        def __init__(self,input_features=2048,n_pos=19,n_limbs=19,quad_size=2,hout=46,wout=46,stride=8,mesh_grid=None,data_format="channels_first"):
             super().__init__()
             self.input_features=input_features
             self.n_pos=n_pos
             self.n_limbs=n_limbs
             self.quad_size=quad_size
+            self.hout=hout
+            self.wout=wout
+            self.stride=stride
             self.out_features=self.n_limbs*9*(self.quad_size**2)
+            self.mesh_grid=mesh_grid
             self.data_format=data_format
             self.tf_data_format="NCHW" if self.data_format=="channels_first" else "NHWC"
             self.main_block=Conv2d(n_filter=self.out_features,in_channels=self.input_features,filter_size=(1,1),data_format=self.data_format)
@@ -176,7 +191,7 @@ class Pifpaf(Model):
         def forward(self,x,is_train=False):
             x=self.main_block.forward(x)
             x=tf.nn.depth_to_space(x,block_size=self.quad_size,data_format=self.tf_data_format)
-            x=tf.reshape(x,[x.shape[0],self.n_limbs,9,x.shape[2],x.shape[3]])
+            x=tf.reshape(x,[-1,self.n_limbs,9,self.hout,self.wout])
             paf_conf=x[:,:,0,:,:]
             paf_src_vec=x[:,:,1:3,:,:]
             paf_dst_vec=x[:,:,3:5,:,:]
@@ -186,7 +201,10 @@ class Pifpaf(Model):
             paf_dst_scale=tf.exp(x[:,:,8,:,:])
             #restore vec_maps in inference
             if(is_train==False):
-                paf_conf=tf.nn.sigmoid(paf_conf)
-                paf_src_vec[:,:]+=self.mesh_grid
-                paf_dst_vec[:,:]+=self.mesh_grid
+                infer_paf_conf=tf.nn.sigmoid(paf_conf)
+                infer_paf_src_vec=(paf_src_vec[:,:]+self.mesh_grid)*self.stride
+                infer_paf_dst_vec=(paf_dst_vec[:,:]+self.mesh_grid)*self.stride
+                infer_paf_src_scale=paf_src_scale*self.stride
+                infer_paf_dst_scale=paf_dst_scale*self.stride
+                return infer_paf_conf,infer_paf_src_vec,infer_paf_dst_vec,paf_src_logb,paf_dst_logb,infer_paf_src_scale,infer_paf_dst_scale
             return paf_conf,paf_src_vec,paf_dst_vec,paf_src_logb,paf_dst_logb,paf_src_scale,paf_dst_scale
